Move Slack URL checking to chat/index
Previously checked for and removed Slack URL objects only in certain
fields and just before sending the ticket to Freshdesk. But users can
put URLs into other fields, too, and they can be buried among other
text. So moved the functionality to index in order to check all input
for URLs, and strip out the Slack URL objects BEFORE saving to the
ticket in progress.